+++ examples/HRDNS/System/Network/Sniffer.php
@@ -14,7 +14,7 @@ class Sniffer extends Command
 
     protected function configure()
     {
-        $this->setName('exmaple:' . strtolower(str_replace('.php', '', basename(__FILE__))));
+        $this->setName('example:' . strtolower(str_replace('.php', '', basename(__FILE__))));
         $this->setProcessTitle($this->getName());
         $this->setDescription('A PHP incoming network sniffer.');
         $this->setHelp('');