+++ src/HRDNS/HomeMatic/BinRpcDecoder.php
@@ -2,6 +2,8 @@
 
 namespace HRDNS\HomeMatic;
 
+use HRDNS\Exception\HomeMaticIOException;
+
 class BinRpcDecoder
 {
 
@@ -13,7 +15,7 @@ class BinRpcDecoder
     public function decode(string $data): array
     {
         if (strlen($data) < 4) {
-            throw new \InvalidArgumentException('Argument 1 is to short for a homematic binrpc packet.', 1);
+            throw new \InvalidArgumentException('Argument 1 is to short for a homematic binrpc packet. (1)', 1);
         }
         $packet = unpack('A3prefix/Ctype', $data);
         if ($packet['prefix'] !== BinRpcProtocol::PREFIX) {
@@ -24,6 +26,10 @@ class BinRpcDecoder
         }
         if ($packet['type'] === BinRpcProtocol::TYPE_RESPONSE) {
             return $this->decodeResponse($data);
+        }
+        if ($packet['type'] === BinRpcProtocol::TYPE_ERROR) {
+            $error = $this->decodeError($data);
+            throw new HomeMaticIOException($error['params']['faultString'], (int)$error['params']['faultCode']);
         }
         throw new \InvalidArgumentException('Argument 1 is an invalid homematic binrpc message type.', 3);
     }
@@ -36,7 +42,7 @@ class BinRpcDecoder
     public function decodeRequest(string $data): array
     {
         if (strlen($data) < 12) {
-            throw new \InvalidArgumentException('Argument 1 is to short for a homematic binrpc packet.', 1);
+            throw new \InvalidArgumentException('Argument 1 is to short for a homematic binrpc packet. (2)', 4);
         }
         $format = 'A3prefix/Ctype/NmsgSize/NmethodSize';
         $packet = unpack($format, $data);
@@ -46,15 +52,24 @@ class BinRpcDecoder
         if ($packet['type'] !== BinRpcProtocol::TYPE_REQUEST) {
             throw new \InvalidArgumentException('Argument 1 is not a homematic binrpc request.', 6);
         }
-        $format .= '/A' . $packet['methodSize'] . 'methodName';
+        $format .= '/A' . $packet['methodSize'] . 'methodName/Ncount';
         $packet = unpack($format, $data);
 
+        if (strlen($data)-8 != $packet['msgSize']) {
+            throw new HomeMaticIOException('Not enough data.',HomeMaticIOException::ERROR_PARSER_NOT_ENOUGH_INPUT);
+        }
+
         $data = substr($data, 3 + 1 + 4 + 4 + $packet['methodSize'] + 4);
+
+        $params = [];
+        for ($i = 0; $i < $packet['count']; $i++) {
+            $params[] = $this->decodeData($data);
+        }
 
         return [
             'type' => 'request',
             'method' => $packet['methodName'],
-            'params' => $this->decodeData($data)
+            'params' => $params,
         ];
     }
 
@@ -66,15 +81,19 @@ class BinRpcDecoder
     public function decodeResponse(string $data): array
     {
         if (strlen($data) < 8) {
-            throw new \InvalidArgumentException('Argument 1 is to short for a homematic binrpc packet.', 1);
+            throw new \InvalidArgumentException('Argument 1 is to short for a homematic binrpc packet. (3)', 7);
         }
         $format = 'A3prefix/Ctype/NmsgSize/A*data';
         $packet = unpack($format, $data);
         if ($packet['prefix'] !== BinRpcProtocol::PREFIX) {
             throw new \InvalidArgumentException('Argument 1 is not a homematic binrpc string.', 8);
         }
-        if ($packet['type'] !== BinRpcProtocol::TYPE_RESPONSE) {
+        if ($packet['type'] !== BinRpcProtocol::TYPE_RESPONSE && $packet['type'] !== BinRpcProtocol::TYPE_ERROR) {
             throw new \InvalidArgumentException('Argument 1 is not a homematic binrpc response.', 9);
+        }
+
+        if (strlen($data)-8 != $packet['msgSize']) {
+            throw new HomeMaticIOException('Not enough data.',HomeMaticIOException::ERROR_PARSER_NOT_ENOUGH_INPUT);
         }
 
         $data = substr($data, 3 + 1 + 4);
@@ -89,26 +108,46 @@ class BinRpcDecoder
     /**
      * @param string $data
      * @return mixed
+     */
+    public function decodeError(string $data): array
+    {
+        $result = $this->decodeResponse($data);
+        $result['type'] = 'error';
+
+        return $result;
+    }
+
+    /**
+     * @param string $data
+     * @return mixed
      * @throws \InvalidArgumentException
      */
     private function decodeData(string &$data)
     {
-        $info = unpack('Ntype', $data);
-        switch ($info['type']) {
-            case BinRpcProtocol::TYPE_INTEGER:
-                return $this->decodeInteger($data);
-            case BinRpcProtocol::TYPE_BOOL:
-                return $this->decodeBool($data);
-            case BinRpcProtocol::TYPE_STRING:
-                return $this->decodeString($data);
-            case BinRpcProtocol::TYPE_FLOAT:
-                return $this->decodeFloat($data);
-            case BinRpcProtocol::TYPE_STRUCT:
-                return $this->decodeStruct($data);
-            case BinRpcProtocol::TYPE_ARRAY:
-                return $this->decodeArray($data);
+        if (!strlen($data)) {
+            throw new \InvalidArgumentException('Argument 1 could not be empty.', 10);
         }
-        throw new \InvalidArgumentException('Invalid argument 1, unable to convert from binrpc format.', 10);
+        $info = unpack('Ntype', $data);
+        if (is_array($info) && array_key_exists('type', $info)) {
+            switch ($info['type']) {
+                case BinRpcProtocol::TYPE_INTEGER:
+                    return $this->decodeInteger($data);
+                case BinRpcProtocol::TYPE_BOOL:
+                    return $this->decodeBool($data);
+                case BinRpcProtocol::TYPE_STRING:
+                    return $this->decodeString($data);
+                case BinRpcProtocol::TYPE_FLOAT:
+                    return $this->decodeFloat($data);
+                case BinRpcProtocol::TYPE_ARRAY:
+                    return $this->decodeArray($data);
+                case BinRpcProtocol::TYPE_STRUCT:
+                    return $this->decodeStruct($data);
+                default:
+                    throw new \InvalidArgumentException('Invalid type ' . $info['type'] . ' found in binrpc request.',
+                        11);
+            }
+        }
+        throw new \InvalidArgumentException('Invalid argument 1. unable to convert from binrpc format.', 12);
     }
 
     /**
@@ -117,11 +156,10 @@ class BinRpcDecoder
      */
     private function decodeFloat(string &$data): float
     {
-        $info = unpack('Ntype/lmantissa/lexponent', $data);
-        $result = round((pow(2, $info['exponent'])) * ($info['mantissa'] / (1 << 30)), 6);
+        $info = unpack('Ntype/Nmantissa/Nexponent', $data);
         $data = substr($data, 4 + 4 + 4);
 
-        return $result;
+        return round((pow(2, $info['exponent'])) * ($info['mantissa'] / (1 << 30)),6);
     }
 
     /**
@@ -130,11 +168,10 @@ class BinRpcDecoder
      */
     private function decodeInteger(string &$data): int
     {
-        $info = @unpack('Ntype/lvalue', $data);
-        $result = (int)$info['value'];
+        $info = unpack('Ntype/Nvalue', $data);
         $data = substr($data, 4 + 4);
 
-        return $result;
+        return (int)$info['value'];
     }
 
     /**
@@ -144,10 +181,9 @@ class BinRpcDecoder
     private function decodeBool(string &$data): bool
     {
         $info = unpack('Ntype/Cvalue', $data);
-        $result = (bool)$info['value'];
         $data = substr($data, 4 + 1);
 
-        return $result;
+        return (bool)$info['value'];
     }
 
     /**
@@ -158,10 +194,9 @@ class BinRpcDecoder
     {
         $info = unpack('Ntype/Nsize', $data);
         $info = unpack('Ntype/Nsize/A' . $info['size'] . 'content', $data);
-        $result = (string)$info['content'];
         $data = substr($data, 4 + 4 + $info['size']);
 
-        return $result;
+        return (string)$info['content'];
     }
 
     /**